requirement.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-camel-case==1.2.0
Markdown==3.3.4
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
